Refactor weights services
